pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-pwa-assets"
version = "0.2.0"
description = "Complete PWA asset generation for Django: icons, splash screens, favicons, Windows tiles, dark mode. Inspired by pwa-asset-generator."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
authors = [{ name = "Maurizio Melani", email = "maurizio.melani@gmail.com" }]
keywords = ["django", "pwa", "progressive-web-app", "icons", "splash-screen", "favicon", "manifest"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
]

dependencies = [
    "django>=3.2",
    "Pillow>=7.0",
]

[project.optional-dependencies]
blank3 = ["blake3>=1.0"]
svg = ["cairosvg>=2.7", "scour>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "cairosvg>=2.7",
    "scour>=0.38.1",
    "blake3>=1.0",
]

[project.urls]
Homepage      = "https://github.com/forwalk-org/django_pwa_assets"
Documentation = "https://github.com/forwalk-org/django_pwa_assets#readme"
"Bug Tracker" = "https://github.com/forwalk-org/django_pwa_assets/issues"
Changelog     = "https://github.com/forwalk-org/django_pwa_assets/blob/main/CHANGELOG.md"

# ── pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files  = ["test_*.py"]
addopts       = "--tb=short -q"

# ── coverage ────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["django_pwa_assets"]
omit   = [
    "*/tests/*",
    "*/migrations/*",
    "*/management/commands/*",
]

[tool.coverage.report]
show_missing = true
fail_under   = 65
